Revert "Refactored mock_tests/ folder"
This reverts commit 06edeb9b2bfbc3e07b4a7f6f05aa6530336f0998.
