+++ tests/mock_tests/mock_hiredis.cpp
